add type casting

+++ tests/tools/TestApp/src/json_context.cc
@@ -6,8 +6,8 @@ namespace mesh {
     MeshConfig_Video get_video_params(MeshConnection *conn){
       ConnectionContext *conn_ctx = (ConnectionContext *)conn;
       return (MeshConfig_Video){
-        .width = conn_ctx->cfg_json.payload.video.width,
-        .height = conn_ctx->cfg_json.payload.video.height,
+        .width = (int)conn_ctx->cfg_json.payload.video.width,
+        .height = (int)conn_ctx->cfg_json.payload.video.height,
         .fps = conn_ctx->cfg_json.payload.video.fps,
         .pixel_format = conn_ctx->cfg_json.payload.video.pixel_format
       };
